% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnspop.R
\name{bootstrap_population_estimate}
\alias{bootstrap_population_estimate}
\title{Estimate population size given privatized (hashed) neighbor identifiers with bootstrap confidence intervals}
\usage{
bootstrap_population_estimate(
  subject,
  recruiter,
  subject_hash,
  degree,
  nbrs,
  rho = NULL,
  n_bootstrap = 500,
  conf_level = 0.95,
  progress = TRUE
)
}
\arguments{
\item{subject}{The integer ids of each subject}

\item{recruiter}{The integer ids of the recruiter of each subject (-1 for seeds)}

\item{subject_hash}{The hashed identifier for the subject}

\item{degree}{The degree of each subject}

\item{nbrs}{A list, each element indicating the hashed identifier of the neighbors of each subject,
or a random subset of those neighbors.}

\item{rho}{The probability two random individuals have the same hash value.
If NULL this is estimated from the number of hash collisions in subject_hash.}

\item{n_bootstrap}{The number of bootrap samples used to estimate the confidence intervals}

\item{conf_level}{The coverage level for the confidence intervals}

\item{progress}{Show a text progress bar}
}
\value{
A data frame with the population size values, CI lower and CI upper bounds.
}
\description{
Estimate population size given privatized (hashed) neighbor identifiers with bootstrap confidence intervals
}
\examples{
data(faux_pns)
#rho estimated from data


#hashes
population_estimate_hash(rds$subject,rds$recruiter,
     subj_hash, rds$degree, nbrs_hash, rho)
\dontrun{
bootstrap_population_estimate(rds$subject,rds$recruiter,
  subj_hash, rds$degree, nbrs_hash, rho, 50)
}
}
