% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnspop.R
\name{cross_tree_pse}
\alias{cross_tree_pse}
\title{Estimate population size given privatized (hashed) neighbor identifiers}
\usage{
cross_tree_pse(
  subject,
  recruiter,
  subject_hash,
  degree,
  nbrs,
  rho = NULL,
  method = c("network", "alter", "sample"),
  small_sample_fraction = TRUE
)
}
\arguments{
\item{subject}{The integer ids of each subject}

\item{recruiter}{The integer ids of the recruiter of each subject (-1 for seeds)}

\item{subject_hash}{The hashed identifier for the subject}

\item{degree}{The degree of each subject}

\item{nbrs}{A list, each element indicating the hashed identifier of the neighbors of each subject,
or a random subset of those neighbors.}

\item{rho}{The probability two random individuals have the same hash value. If NULL this is estimated from the number of hash collisions in subject_hash.}

\item{method}{combined uses both the sample and the nominated alters for potential matches. "sample" uses only recruited individuals and "alter" uses only their nominated alters.}

\item{small_sample_fraction}{If TRUE, simplifies estimation by assuming a small sample fraction.}
}
\value{
A data.frame with elements:
'estimate': The population size estimate, adjusted for hashing
'rho': The hash collision probability
'num_matches': The number of identifier matches
'num_poss_matches': The number of identifier match opportunities
}
\description{
Estimate population size given privatized (hashed) neighbor identifiers
}
\examples{
data(faux_pns)
#rho estimated from data
cross_tree_pse(faux_pns$subject,faux_pns$recruiter,
               faux_pns$subject_hash, faux_pns$degree,
               faux_pns[paste0("friend_hash",1:11)])


# fixed rho
cross_tree_pse(faux_pns$subject,faux_pns$recruiter,
               faux_pns$subject_hash, faux_pns$degree,
               faux_pns[paste0("friend_hash",1:11)], rho=.001)
}
