% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnspop.R
\docType{data}
\name{faux_pns}
\alias{faux_pns}
\alias{rds}
\alias{nbrs_hash}
\alias{nbrs2}
\alias{rho}
\alias{subj_hash}
\title{This is example pns data}
\description{
This is example pns data
}
\examples{
data(faux_pns)
#rho estimated from data


#hashes
cross_tree_pse(faux_pns$subject,faux_pns$recruiter,
     faux_pns$subject_hash, faux_pns$degree,
     faux_pns[paste0("friend_hash",1:11)],
     rho=.001)
}
\keyword{data}
