% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnspop.R
\name{getSeed}
\alias{getSeed}
\title{get the root seed for each subject}
\usage{
getSeed(id, recruiter.id)
}
\arguments{
\item{id}{subject ids}

\item{recruiter.id}{subject recruiter id}
}
\description{
get the root seed for each subject
}
