% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnspop.R
\name{makeConfigurationGraph}
\alias{makeConfigurationGraph}
\title{Constructs a configuration graph}
\usage{
makeConfigurationGraph(d)
}
\arguments{
\item{d}{a vector of degrees for each node}
}
\value{
A matrix with 3 columns and length(d) rows,
with each row representing an edge. Edges are from
the first column to the second column, with the third column
indicating whether the edge is reciprocated (1 -> reciprocated, 0 -> directed).
}
\description{
Constructs a configuration graph
}
\examples{
set.seed(1)
makeConfigurationGraph(c(2,3,4,5))
}
\references{
Michael S. Molloy and Bruce A. Reed. A critical point for random graphs with a
given degree sequence. Random Structures and Algorithms, 6:161–179, 1995.
}
