% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnspop.R
\name{overlap_statistics}
\alias{overlap_statistics}
\title{Calculates relevant summary statistics for PNS}
\usage{
overlap_statistics(subject_hash, nbrs)
}
\arguments{
\item{subject_hash}{The hashed identifier for the subject}

\item{nbrs}{A list, each element indicating the hashed identifier of the neighbors of each subject,
or a random subset of those neighbors.}
}
\value{
A list with elements:
   'unique' - 'unique_nbrs' The total number of unique hash values in the nominations.
   'unique' - 'sample_size' The sample size of the PNS
   'unique' - 'total_unique_ident' The unique nominations plus the number of subjects whose hashes
   do not appear in the nominations. This represents the total number of hashed identities directly
   identified by sampling (either through recruitment or nomination) and is thus a hard lower bound
   on population size.

   'neighbors' - 'total_nbrs' The number of nominations
   'neighbors' - 'unique_nbrs' The number of unique nominations

   'naive_crc_estimate' - 'unique_nbrs' The number of unique nominations
   'naive_crc_estimate' - 'sample_size' The sample size of the PNS
   'naive_crc_estimate' - 'unique_nbrs_sample_overlap' The number of nominations that match a hash in the sample
   'naive_crc_estimate' - 'N' Treating the sample and the nominations as two independent capture events,
   this is the population size estimate. This should be treated as a descriptive statistic and not a true
   estimate as the assumption does not hold.
}
\description{
Calculates relevant summary statistics for PNS
}
\examples{
data(faux_pns)

overlap_statistics(faux_pns$subject, faux_pns[paste0("friend_hash",1:11)])
}
