% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnspop.R
\name{populationEstimate}
\alias{populationEstimate}
\title{Estimate population size given neighbor identifiers}
\usage{
populationEstimate(subject, recruiter, degree, nbrs)
}
\arguments{
\item{subject}{The integer ids of each subject}

\item{recruiter}{The integer ids of the recruiter of each subject}

\item{degree}{The degree of each subject}

\item{nbrs}{A list, each element indicating the ids of the neighbors of each subject,
or a random subset of those neighbors}
}
\value{
A list with elements:
'noDegreeEstimate': The nieve n_1 capture re-capture estimate
'estimate': The n_2 population size estimate
'crossSeedEstimate' : The n_3 cross-seed estimate
}
\description{
Estimate population size given neighbor identifiers
}
\examples{
set.seed(1)
n <- 1000 #pop size
d <- rpois(n,lambda = 3) + 1
g <- rep(1,n)
el <- makeConfigurationGraph(d)
seeds <- 7 # #of seeds
rds <- sampRDS(el, d, seeds,g,200,FALSE, pr = c(0,.1,.9))
nbrs <- lapply(1:n, function(i) c(el[el[,1]==i,2],el[el[,2]==i,1]))
nbrs2 <- nbrs[rds$subject]
populationEstimate(rds$subject,rds$recruiter, d[rds$subject], nbrs2)
}
\references{
Khan, Bilal; Lee, Hsuan-Wei; Fellows, Ian; Dombrowski, Kirk One-step Estimation of Networked Population Size: Respondent-Driven Capture-Recapture with Anonymity eprint arXiv:1710.03953, 2017
}
