% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnspop.R
\name{sampRDS}
\alias{sampRDS}
\title{Create and RDS sample from a graph}
\usage{
sampRDS(el, d, ns, g, ss, biased = TRUE, pr = c(0, 0.1, 0.9))
}
\arguments{
\item{el}{A matrix with three columns, where each row is an edge. Edges are from
the first column to the second column, with the third column
indicating whether the edge is reciprocated (1 -> reciprocated, 0 -> directed).}

\item{d}{The degrees of each node}

\item{ns}{The number of seeds}

\item{g}{A factor variable for use with biased sampling}

\item{ss}{The size of the RDS sample}

\item{biased}{If true, all seeds are drawn from subjects with the highest level of g.}

\item{pr}{A probability vector for the number of recruits. The first element is the
probability that a subject tries to recruit no neighbors. The second element is the probability
that a subject tries to recruit 1 neightbor, and so on. Subjects can not recruit more
neighbors than they have available.}
}
\value{
A data.frame with columns:
'subject' The index the recruited subject
'recruiter' The index of the recruiter of the subject (-1 if seed)
'time' the time of the subject's recruitment
}
\description{
Create and RDS sample from a graph
}
\details{
Subjects recruit neighbors at random, with a time delay of rexp(1). Sampling stops
when ss is reached. If no recruit is available, a new seed is drawn and 'redraw' is
printed to the console.
}
\examples{
set.seed(1)
n <- 1000 #pop size
d <- rpois(n,lambda = 3) + 1
g <- rep(1,n)
el <- makeConfigurationGraph(d)
seeds <- 7 # #of seeds
rds <- sampRDS(el, d, seeds,g,200,FALSE, pr = c(0,.1,.9))
}
