% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnspop.R
\name{shiny_pnspop}
\alias{shiny_pnspop}
\title{Launches the Shiny PNS Application}
\usage{
shiny_pnspop(...)
}
\arguments{
\item{...}{Additional parameters to be passed to shiny::runApp}
}
\description{
Launches the Shiny PNS Application
}
