% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnspop.R
\docType{data}
\name{faux_pns}
\alias{faux_pns}
\alias{rds}
\alias{nbrs_hash}
\alias{nbrs2}
\alias{rho}
\alias{subj_hash}
\title{This is example pns data}
\description{
This is example pns data
}
\examples{
data(faux_pns)
#rho estimated from data

# no hashes
population_estimate_hash(rds$subject,rds$recruiter, rds$subject, rds$degree, nbrs2, rho=0)

#hashes
population_estimate_hash(rds$subject,rds$recruiter,
     subj_hash, rds$degree, nbrs_hash)
}
\keyword{data}
